+++ docs/module.rst
@@ -8,3 +8,19 @@
     :special-members:
     :show-inheritance:
     :exclude-members: __dict__, __hash__, __module__, __weakref__, __enter__, __exit__, __annotations__
+
+.. automodule:: filecache.file_cache_prefix
+    :member-order: bysource
+    :members:
+    :undoc-members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __dict__, __hash__, __module__, __weakref__, __enter__, __exit__, __annotations__
+
+.. automodule:: filecache.file_cache_source
+    :member-order: bysource
+    :members:
+    :undoc-members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __dict__, __hash__, __module__, __weakref__, __enter__, __exit__, __annotations__
